+++ src/filters/boolean_filter.ts
@@ -152,7 +152,7 @@ export default class BooleanFilter extends TootFilter {
             (a, b) => compareStr(a.displayName || a.name, b.displayName || b.name)
         );
 
-        return this.optionListWithMinToots(options, minToots, includeFollowed).filter(obj => (obj.numToots || 0) >= 1);
+        return this.optionListWithMinToots(options, minToots, includeFollowed);
     }
 
     /**
@@ -222,7 +222,8 @@ export default class BooleanFilter extends TootFilter {
         includeFollowed: boolean = false
     ): BooleanFilterOptionList {
         const newOptions = options.filter(o => {
-            return (o.numToots || 0) >= minToots || this.isOptionEnabled(o.name) || (includeFollowed && o.isFollowed)
+            const numToots = o.numToots || 0;
+            return numToots >= minToots || this.isOptionEnabled(o.name) || (includeFollowed && o.isFollowed && numToots > 0)
         });
 
         this.selectedOptions.forEach((selected) => {